// reactapp/src/store/inputValues/inputValues.type.ts
export enum INPUT_VALUES_ACTION_TYPES {
  SET_TEAM = "input_values/SET_TEAM",
  SET_FORMATION = "input_values/SET_FORMATION",
  SET_DIRECTION = "input_values/SET_DIRECTION",
  SET_PLAYER = "input_values/SET_PLAYER",
}

export type Info = {
  // value: string | number;
  value: string;
  name: string;
};